Suppressed large task graph warning caused by some datasets
This might be a performance issue, but the fixes I was able to develop using client.scatter caused major memory issues as one worker attempts to concatenate the whole dataset in memory before writing to disk. This causes slow downs on virtually every other dataset that doesnt cause the warning, so at the moment this fix is not worth it.